CA destruction, CA custom keysize
CA destruction command
========================
This command allows the user to destroy his
CA with `./caman destroy`. The user needs to confirm its choice by
writing "Yes, I understand the risks" when prompted, because this
operation destroys the whole CA dir, and the saved certificates.

Custom private key size for the CA
=====================================
The user can specify in the caconfig.cnf file the size of the CA key,
by specifying it in "default_bits". I personally like to have a
strong security for my CAs, therefore I change my default_bits to 8192.